requirements.txt
llama-index
llama-index-core
llama-cloud
llama-index-llms-openai
llama-parse
python-dotenv
openai